requirements.txt
asyncua==1.1.0
numpy==1.26.4
opcua==0.98.13
pandas==2.2.2
plotly==5.21.0
python-dotenv==1.0.1
streamlit==1.33.0